ch04 softdelete clear!

+++ resources/js/brand_delete.js
@@ -1,12 +1,24 @@
 let token = document.querySelector('meta[name="csrf-token"]').getAttribute('content');
-delete_brand();
+
+choice();
+function choice(){
+    let choice_value = confirm('삭제 하시겠습니까?');
+    if(choice_value == true) {
+        delete_brand();
+    }
+    if(choice_value == false) {
+        alert('취소되었습니다.');
+        location.href('./brand');
+    }
+} 
 
 async function delete_brand() {
     let idx_value = sessionStorage.getItem('brand_change');
     console.log(idx_value);
-    
-    let url = "brand_delete_serve"
-    return fetch(url, {
+    // console.log(cookie);
+    let url = "brand_delete_serve";
+
+    fetch(url, {
         headers: {
             "Content-Type": "application/json",
             "Accept": "application/json, text-plain, */*",
@@ -21,7 +33,8 @@ async function delete_brand() {
     })
     .then(response => response.json())
     .then(response => {
-        return response;
+        alert('삭제성공!');
+        location.href('./brand');
     })
     .catch(function(error) {
         console.log(error);
